+++ src/hooks/useAuthenticate.js
@@ -1,15 +1,14 @@
 import { useState } from "react";
-import useAuthContext from "./useAuthContext";
 import useAxios from "./useAxios";
 import { useNavigate } from "react-router-dom";
+import useAuthStore from "../store/authStore";
 
 export default function useAuthenticate() {
     const [error, setError] = useState("");
     const [isLoading, setIsLoading] = useState(false);
     const axios = useAxios();
     const navigate = useNavigate();
-
-    const { setAuth } = useAuthContext();
+    const { setAuth } = useAuthStore();
 
     const authenticate = async (mode, authData) => {
         setError("");